backend/requirements.txt
django==2.2.16
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
django-filter==2.4.0
Pillow==8.3.1
sorl-thumbnail==12.7.0
drf-base64==2.0
django-colorfield
reportlab==3.6.6
isort==5.10.1